Update avidemux_fr.ts
Hi
 I have done the update of one line of the Avidemux_fr.ts file.

For the rest I am in the process of finishing the complete translation of the "avidemux_fr.ts" file on the missing terms and the errors of
interpretation or poorly adapted wording.

I searched everywhere on my PC where the "avidemux_fr.xml" or .ts file could be found without finding anything. 
I am therefore having difficulty validating all my proposals because I cannot reproduce the action of my changes.

It may indeed happen that my proposals are too long depending on the available space or are not adapted to the correct screen or dialog box. 
I am therefore interested in a way to test my work without necessarily going through the source code, the implementation of which I do not understand.

Thank you for your feedback and thank you to the developers of
Avidemux who offer free software that is very useful for anyone who wants to adapt videos to their hardware environment.

best regard